[fix] JPA 리포지토리, 엔티티 간 PK 타입 수정

+++ src/main/java/site/offload/offloadserver/db/repository/GainedCharacterRepository.java
@@ -3,5 +3,5 @@ package site.offload.offloadserver.db.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import site.offload.offloadserver.db.entity.GainedCharacter;
 
-public interface GainedCharacterRepository extends JpaRepository<GainedCharacter, Integer> {
+public interface GainedCharacterRepository extends JpaRepository<GainedCharacter, Long> {
 }
